+++ app/keepers.go
@@ -335,7 +335,7 @@ func (app *OsmosisApp) InitNormalKeepers(
 	app.SuperfluidKeeper = superfluidkeeper.NewKeeper(
 		appCodec, keys[superfluidtypes.StoreKey], app.GetSubspace(superfluidtypes.ModuleName),
 		*app.AccountKeeper, app.BankKeeper, app.StakingKeeper, app.DistrKeeper, app.EpochsKeeper, app.LockupKeeper, gammKeeper, app.IncentivesKeeper,
-		lockupkeeper.NewMsgServerImpl(app.LockupKeeper))
+		lockupkeeper.LockKeeperImpl(app.LockupKeeper))
 
 	mintKeeper := mintkeeper.NewKeeper(
 		appCodec, keys[minttypes.StoreKey],
