+++ x/lockup/keeper/keeper.go
@@ -125,3 +125,88 @@ func (server msgServer) LockTokens(goCtx context.Context, msg *types.MsgLockToke
 
 	return &types.MsgLockTokensResponse{ID: lock.ID}, nil
 }
+
+type msgServer struct {
+	keeper *Keeper
+}
+
+// LockKeeper is the server API for Msg service.
+type LockKeeper interface {
+	// LockTokens lock tokens
+	LockTokens(context.Context, *types.MsgLockTokens) (*types.MsgLockTokensResponse, error)
+	// BeginUnlockingAll begin unlocking all tokens
+	BeginUnlockingAll(context.Context, *types.MsgBeginUnlockingAll) (*types.MsgBeginUnlockingAllResponse, error)
+	// MsgBeginUnlocking begins unlocking tokens by lock ID
+	BeginUnlocking(context.Context, *types.MsgBeginUnlocking) (*types.MsgBeginUnlockingResponse, error)
+}
+
+func (server msgServer) BeginUnlocking(goCtx context.Context, msg *types.MsgBeginUnlocking) (*types.MsgBeginUnlockingResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	lock, err := server.keeper.GetLockByID(ctx, msg.ID)
+	if err != nil {
+		return nil, errors.Wrap(errors.ErrInvalidRequest, err.Error())
+	}
+
+	err = server.keeper.BeginUnlock(ctx, lock.ID, msg.Coins)
+	if err != nil {
+		return nil, errors.Wrap(errors.ErrInvalidRequest, err.Error())
+	}
+
+	if msg.Owner != lock.Owner {
+		return nil, errors.Wrap(types.ErrNotLockOwner, fmt.Sprintf("msg sender(%s) and lock owner(%s) does not match", msg.Owner, lock.Owner))
+	}
+
+	ctx.EventManager().EmitEvents(sdk.Events{
+		createBeginUnlockEvent(lock),
+	})
+
+	return &types.MsgBeginUnlockingResponse{}, nil
+}
+
+func (server msgServer) BeginUnlockingAll(goCtx context.Context, msg *types.MsgBeginUnlockingAll) (*types.MsgBeginUnlockingAllResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	owner, err := sdk.AccAddressFromBech32(msg.Owner)
+	if err != nil {
+		return nil, err
+	}
+
+	unlocks, err := server.keeper.BeginUnlockAllNotUnlockings(ctx, owner)
+	if err != nil {
+		return nil, errors.Wrap(errors.ErrInvalidRequest, err.Error())
+	}
+
+	// Create the events for this message
+	unlockedCoins := server.keeper.getCoinsFromLocks(unlocks)
+	events := sdk.Events{
+		sdk.NewEvent(
+			types.TypeEvtBeginUnlockAll,
+			sdk.NewAttribute(types.AttributePeriodLockOwner, msg.Owner),
+			sdk.NewAttribute(types.AttributeUnlockedCoins, unlockedCoins.String()),
+		),
+	}
+	for _, lock := range unlocks {
+		events = events.AppendEvent(createBeginUnlockEvent(&lock))
+	}
+	ctx.EventManager().EmitEvents(events)
+
+	return &types.MsgBeginUnlockingAllResponse{}, nil
+}
+
+func createBeginUnlockEvent(lock *types.PeriodLock) sdk.Event {
+	return sdk.NewEvent(
+		types.TypeEvtBeginUnlock,
+		sdk.NewAttribute(types.AttributePeriodLockID, utils.Uint64ToString(lock.ID)),
+		sdk.NewAttribute(types.AttributePeriodLockOwner, lock.Owner),
+		sdk.NewAttribute(types.AttributePeriodLockDuration, lock.Duration.String()),
+		sdk.NewAttribute(types.AttributePeriodLockUnlockTime, lock.EndTime.String()),
+	)
+}
+
+// LockKeeperImpl returns an instance of LockKeeper
+func LockKeeperImpl(keeper *Keeper) LockKeeper {
+	return &msgServer{
+		keeper: keeper,
+	}
+}
